* [android] migrate drag feature into BasicListComponent
